Powerplant B2 (Zapdos) - Non-switch-puzzle version
Made the basement for Zapdos. This is adapted from the original concept,
will make tiles if we want to do the puzzle like originally intended,
but for now, this works.

Someone should group the maps better than me leaving them default. Not
sure how to properly and don't want to break things.